pyproject.toml
[project]
name = "brainiac-2"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "asyncio>=3.4.3",
  "click>=8.1.8",
  "configparser>=7.2.0",
  "openai>=1.76.2",
  "pydantic>=2.11.4",
  "python-slugify>=8.0.4",
  "readtime>=3.0.0",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project.scripts]
brainiac-2 = "brainiac.main:cli"
